# Don't allocate slice to pick candidate steps.